+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 - Require Mopidy v1.1 or later.
 
+- Use ``Extension.get_data_dir()`` as default image directory.
+
 
 0.3.3 2015-06-17
 ----------------